+++ src/main/java/com/project/pescueshop/model/entity/Variety.java
@@ -35,10 +35,10 @@ public class Variety {
     )
     private List<VarietyAttribute> varietyAttributes;
     private String status;
-    private Double width = 0.0d;
-    private Double height = 0.0d;
-    private Double length = 0.0d;
-    private Double weight = 0.0d;
+    private Integer width = 0;
+    private Integer height = 0;
+    private Integer length = 0;
+    private Integer weight = 0;
     private Integer stockAmount = 0;
 
     public Variety(VarietyDTO dto){
